Replaced REST API with GraphQL API

+++ src/main/kotlin/com/antonl/cssundays/repositories/PlayerRepository.kt
@@ -3,7 +3,7 @@ package com.antonl.cssundays.repositories
 import com.antonl.cssundays.model.Player
 import org.springframework.data.repository.CrudRepository
 
-interface PlayerRepository : CrudRepository<Player, Long> {
+interface PlayerRepository : CrudRepository<Player, Int> {
     fun findPlayerBySlug(slug: String): Player
-    fun findPlayerById(id: Long): Player
+    fun findPlayerById(id: Int): Player
 }
